Re factirzacion de renewCards. Falta test
